+++ cosmicfishpie/fishermatrix/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-
 """LSS Module
 
 Provide some basic description of the theory module.